// dashboard-ts/src/api-types.d.ts
export type NFCResponseDataType = {
	id: number;
	is_assigned: boolean;
	name: string;
	nfc_data: string;
	timestamp: string;
	expires_at: string;
	device_type: string;
	latitude: number;
	longitude: number;
	altitude: number;
	radius_in_meters: number;
	circle_geojson: {
		type: string;
		coordinates: Array<Array<Array<number>>>;
	};
	created_by: number;
	assigned_to_bandobas: number;
};

export type AddOfficerToBandobastType = {
	id: number;
	police_user: string;
	mobile: string;
	fcm_token: string;
	device_id: string;
	created_by: number;
	assigned_nfc_device: number;
};

export type GetOfficerFromStartTimeToEndtimeType = {
	available_user_ids: Array<{
		id: number;
		name: string;
		phone: string;
		email: string;
		image_url: string;
		rank: string;
		department: string;
	}>;
};